Реализовывает переключение точки маршрута на форму редактирования
