--- src/components/page-main/trip-event-form/photos.js
@@ -1,9 +1,0 @@
-const renderPhotos = (photos) => {
-  return photos.map((photo) => {
-    return (
-      `<img class="event__photo" src="${photo}" alt="Event photo">`.trim());
-  })
-  .join(`\n`);
-};
-
-export {renderPhotos};
