--- src/components/page-main/trip-event-form/render-options.js
@@ -1,8 +1,0 @@
-const renderOptions = (cities) => {
-  return cities.map((city) => {
-    return (`<option value="${city}"></option>`.trim());
-  })
-  .join(`\n`);
-};
-
-export {renderOptions};
