+++ src/components/page-main/trip-events/render-events-in-days.js
@@ -1,7 +1,29 @@
 import TripEventsContainerComponent from '../trip-events/trip-events-container.js';
 import TripDayComponent from '../trip-days/trip-day.js';
 import TripEventComponent from './trip-event.js';
+import TripEventFormComponent from './../trip-event-form/trip-event-form.js';
 import {render} from '../../../helpers/utils.js';
+
+const addTripEventToList = (tripEventListElement, tripEvent) => {
+  const eventRollupButtonClickHandler = () => {
+    tripEventListElement.replaceChild(tripEventEditFormComponent.getElement(), tripEventComponent.getElement());
+  };
+
+  const replaceEditFormToTripEvent = (evt) => {
+    evt.preventDefault();
+    tripEventListElement.replaceChild(tripEventComponent.getElement(), tripEventEditFormComponent.getElement());
+  };
+
+  const tripEventComponent = new TripEventComponent(tripEvent);
+  const eventRollupButton = tripEventComponent.getElement().querySelector(`.event__rollup-btn`);
+  eventRollupButton.addEventListener(`click`, eventRollupButtonClickHandler);
+
+  const tripEventEditFormComponent = new TripEventFormComponent(tripEvent, tripEvent.counter);
+  const eventEditForm = tripEventEditFormComponent.getElement();
+  eventEditForm.addEventListener(`submit`, replaceEditFormToTripEvent);
+
+  render(tripEventListElement, tripEventComponent.getElement());
+};
 
 const renderEventsInDays = (tripEventsSortedByDate, tripDaysObjects, daysContainer) => {
   let daysContainerCount = 0;
@@ -25,7 +47,7 @@ const renderEventsInDays = (tripEventsSortedByDate, tripDaysObjects, daysContain
 
     const tripEventContainer = document.querySelector(`.day:last-child .trip-events__list`);
 
-    render(tripEventContainer, new TripEventComponent(tripEvent).getElement());
+    addTripEventToList(tripEventContainer, tripEvent);
   });
 };
