--- src/components/page-main/trip-events/render-trip-event-offers.js
@@ -1,16 +1,0 @@
-const renderTripEventOffers = (offers) => {
-  return offers.map((offer) => {
-    const {title, price} = offer;
-
-    return (
-      `<li class="event__offer">
-        <span class="event__offer-title">${title}</span>
-        &plus;
-        &euro;&nbsp;<span class="event__offer-price">${price}</span>
-      </li>`.trim()
-    );
-  })
-  .join(`\n`);
-};
-
-export {renderTripEventOffers};
