+++ src/components/page-main/trip-events/trip-event.js
@@ -1,6 +1,20 @@
-import {renderTripEventOffers} from './render-trip-event-offers.js';
 import {getEventTimeFormat} from './trip-event-time.js';
 import {createElement} from '../../../helpers/utils.js';
+
+const renderTripEventOffers = (offers) => {
+  return offers.map((offer) => {
+    const {title, price} = offer;
+
+    return (
+      `<li class="event__offer">
+        <span class="event__offer-title">${title}</span>
+        &plus;
+        &euro;&nbsp;<span class="event__offer-price">${price}</span>
+      </li>`.trim()
+    );
+  })
+  .join(`\n`);
+};
 
 const renderTripEvent = (tripEvent) => {
   const {type, city, basePrice, offers, action, start, end, timeDiff} = tripEvent;
