+++ src/helpers/trip-events-date.js
